+++ qom/ui/axes/MultiAxis.py
@@ -1,12 +1,14 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
  
-"""Class to handle a multi axis."""
+"""Class to handle a multi-value axis."""
 
 __name__    = 'qom.ui.axes.MultiAxis'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-10-10'
-__updated__ = '2020-10-21'
+__updated__ = '2020-12-02'
+
+# TODO: set color and style variants.
 
 # dependencies
 import logging
@@ -19,7 +21,7 @@ from qom.ui.axes.BaseAxis import BaseAxis
 logger = logging.getLogger(__name__)
 
 class MultiAxis(BaseAxis):
-    """Class to handle a multi axis.
+    """Class to handle a multi-value axis.
 
     Inherits :class:`qom.ui.axes.BaseAxis`.
     """
@@ -48,25 +50,29 @@ class MultiAxis(BaseAxis):
         # set legends
         self.legends = ['{name} = {value} {unit}'.format(name=self.name, value=v, unit=self.unit) for v in self.val]
 
-        _dim = len(self.val)
-
         # set colors
         _colors = axis_data.get('colors', [])
-        if type(_colors) is list and len(_colors) == _dim:
+        # if colors are defined
+        if type(_colors) is list and len(_colors) == self.dim:
             self.colors = _colors
+        # else set individual colors
         else:
-            self.colors = ['r' for i in range(_dim)]
+            self.colors = ['r' for i in range(self.dim)]
 
         # set styles
         _styles = axis_data.get('styles', [])
-        if type(_styles) is list and len(_styles) == _dim:
+        # if styles are defined
+        if type(_styles) is list and len(_styles) == self.dim:
             self.styles = _styles
+        # else set individual styles
         else:
-            self.styles = ['-' for i in range(_dim)]
+            self.styles = ['-' for i in range(self.dim)]
 
         # set sizes
         _sizes = axis_data.get('sizes', [])
-        if type(_sizes) is list and len(_sizes) == _dim:
+        # if sizes are defined
+        if type(_sizes) is list and len(_sizes) == self.dim:
             self.sizes = _sizes
+        # else set individual sizes
         else:
-            self.sizes = [2 for i in range(_dim)]
+            self.sizes = [2 for i in range(self.dim)]